Moved files into folders and added Makefiles
