--- Python_Scripts/Calculator.py
@@ -1,82 +1,0 @@
-'''
- * @file Calculator.py
- * @author Oliver Joisten (https://oliver-joisten.se/)
- * @brief A very simple Calculator made in  python's standard library.
-              All operation are done in the terminal.
- * @version 0.1
- * @date 2023-01-12
- *
- * @copyright Copyright (c) 2023
- *
-'''
-
-# This function adds two numbers
-
-
-def add(x, y):
-    return x + y
-
-# This function subtracts two numbers
-
-
-def subtract(x, y):
-    return x - y
-
-# This function multiplies two numbers
-
-
-def multiply(x, y):
-    return x * y
-
-# This function divides two numbers
-
-
-def divide(x, y):
-    return x / y
-
-
-print("Select operation.")
-print("1.Add")
-print("2.Subtract")
-print("3.Multiply")
-print("4.Divide")
-
-while True:
-    # take input from the user
-    choice = input("Enter choice(1/2/3/4): ")
-
-    # check if choice is one of the four options
-    if choice in ('1', '2', '3', '4'):
-        try:
-            num1 = float(input("Enter first number: "))
-            num2 = float(input("Enter second number: "))
-        except ValueError:
-            print("Invalid input. Please enter a number.")
-            continue
-
-        if choice == '1':
-            print(num1, "+", num2, "=", add(num1, num2))
-
-        elif choice == '2':
-            print(num1, "-", num2, "=", subtract(num1, num2))
-
-        elif choice == '3':
-            print(num1, "*", num2, "=", multiply(num1, num2))
-
-        elif choice == '4':
-            print(num1, "/", num2, "=", divide(num1, num2))
-
-        # check if user wants another calculation
-        # break the while loop if answer is no
-        while True:
-            next_calculation = input("Let's do next calculation? (yes/no): ")
-            if next_calculation == "no":
-                break
-            elif next_calculation == "yes":
-                break
-            else:
-                print("Invalid input! Please try again")
-        if next_calculation == "no":
-            break
-    else:
-        print("Invalid Input")
