+++ Python_Scripts/Calculator/Makefile
@@ -1,0 +1,7 @@
+py_script: Calculator.py
+	python3 Calculator.py
+
+app: Calculator.py
+
+.DEFAULT: Calculator.py
+
