+++ Python_Scripts/Calculator_Gui/Makefile
@@ -1,0 +1,6 @@
+py_script: Calculator_gui.py
+	python3 Calculator_gui.py
+
+app: Calculator_gui.py
+
+.DEFAULT: Calculator_gui.py
